Add TMP112 to the schematic and pull-up for TWI
TMP112 is a temperature sensor made by Texas Instruments.

Please note with the pull-up resistors that internally the GPIO pins of the
nRF51 can be configured with an pull-up resistance of ~13K Ohm's so the
external pull-up's might not be necessary.

Also note that any GPIO pin can be selected as a TWI interface, so it is
possible to update the schematic later to ease up the routing part.